Add files via upload
Deep Learning for Text Classification
Process: First some preprocessing of the data was done, dealing with missing values and unbalanced dataset. The dataset was also hot-encoded so that the problem could be approached as multi-class but also as multilabel. We employed a Random Forest as a baseline Machine Learning model, a Neural Network with different configurations and finally a BERT model.
Findings: Baseline ML model gave surprisingly high accuracy given the simplicity of it, a score in kaggle of around 85%. The score was better improved by a Deep Neural Network (91%) which we think is the best model for this dataset. BERT model gave good performance but showed no significant increase as compared to the neural network, and has a much longer training time therefore is not the model we would choose for this dataset.